refactor: split manifest into manifest.go, add missing validation tests
- Move Manifest() and logo embed from github.go into a dedicated manifest.go,
  reducing github.go from 1016 to 147 lines (following Monday connector pattern)
- Add path injection tests to TestCreateOrUpdateFile_PathInjection, mirroring
  the existing tests for get_file_contents
- Add TestListRepos_InvalidParams covering type/visibility/sort enums and per_page limit
- Add TestListPullRequests_InvalidParams covering state/sort/direction enums and per_page
- Fix search_issues sort error message to include valid options (consistency with
  other validation error messages across the connector)
